Release 0.22.0

+++ web-application/client/pages/users/views/UserAccountDangerZoneView.js
@@ -1,5 +1,28 @@
-import React, { useState, useEffect } from 'react'
+/******************************************************************************
+ *
+ *  Communities -- Non-profit, cooperative social media
+ *  Copyright (C) 2022 - 2024 Daniel Bingham
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as published
+ *  by the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ ******************************************************************************/
+import { useState, useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
+
+import logger from '/logger'
+
+import { isLocalStorageAvailable } from '/lib/localStorage'
 
 import { useRequest } from '/lib/hooks/useRequest'
 import { deleteUser } from '/state/User'
@@ -30,9 +53,15 @@ const UserAccountDangerZoneView = function() {
     useEffect(function() {
         if ( request && request.state == 'fulfilled') {
 
-            // Clear local storage so their drafts don't carry over to another
-            // login session.
-            localStorage.clear()
+            if ( isLocalStorageAvailable() ) {
+                try {
+                    // Clear local storage so their drafts don't carry over to another
+                    // login session.
+                    localStorage.clear()
+                } catch (error) {
+                    logger.error(error)
+                }
+            }
 
             dispatch(reset())
 
@@ -54,7 +83,7 @@ const UserAccountDangerZoneView = function() {
                     <div className="danger-zone-view__button-wrapper">
                         <Button type="warn" onClick={(e) => setAreYouSure(true)}>Delete My Account</Button>
                     </div>
-                    <AreYouSure isVisible={areYouSure} execute={deleteCurrentUser} cancel={() => setAreYouSure(false)} > 
+                    <AreYouSure isVisible={areYouSure} execute={deleteCurrentUser} cancel={() => setAreYouSure(false)} >
                         <p>Are you sure you want to delete your account?</p>
                     </AreYouSure>
                 </div>
